+++ include/ghex/transport_layer/ucx/context.hpp
@@ -11,11 +11,13 @@
 #ifndef INCLUDED_TL_MPI_CONTEXT_HPP
 #define INCLUDED_TL_MPI_CONTEXT_HPP
 
-#include "../context.hpp"
-//#include "./communicator.hpp"
-#include "../ucx3/worker.hpp"
-//#include "../common/moved_bit.hpp"
-#include "../../threads/atomic/primitives.hpp"
+//#include "../context.hpp"
+////#include "./communicator.hpp"
+//#include "../ucx3/worker.hpp"
+////#include "../common/moved_bit.hpp"
+//#include "../../threads/atomic/primitives.hpp"
+//#include "./request2.hpp"
+#include "./future2.hpp"
 #include <iostream>
 
 namespace gridtools {
@@ -23,147 +25,6 @@ namespace gridtools {
         namespace tl {
 		    
             namespace ucx {    
-                
-
-                enum class request_kind { send, recv };
-
-                template<typename ThreadPrimitives>
-                struct request_data_ft
-                {
-                    using worker_type            = worker_t<ThreadPrimitives>;
-
-                    void*        m_ucx_ptr;
-                    worker_type* m_recv_worker;
-                    worker_type* m_send_worker;
-                    request_kind m_kind;
-
-                    static constexpr std::uintptr_t mask = ~(alignof(request_data_ft)-1u);
-                    template<typename... Args>
-                    static request_data_ft* construct(void* ptr, Args&& ...args)
-                    {
-                        // align pointer    
-                        auto a_ptr = reinterpret_cast<request_data_ft*>
-                            ((reinterpret_cast<std::uintptr_t>((unsigned char*)ptr) + alignof(request_data_ft)-1) & mask);
-                        new(a_ptr) request_data_ft{ptr,std::forward<Args>(args)...};
-                        return a_ptr;
-                    }
-                };
-                
-                using request_data_size = std::integral_constant<std::size_t,
-                    sizeof(request_data_ft<::gridtools::ghex::threads::atomic::primitives>) +
-                    alignof(request_data_ft<::gridtools::ghex::threads::atomic::primitives>)>;
-                
-                void request_init(void *req) { std::memset(req, 0, request_data_size::value); }
-
-                template<typename ThreadPrimitives>
-                struct request_ft
-                {
-                    using data_type = request_data_ft<ThreadPrimitives>;
-
-                    data_type* m_req = nullptr;
-
-                    request_ft() = default;
-                    request_ft(data_type* ptr) noexcept : m_req{ptr} {}
-                    request_ft(const request_ft&) = delete;
-                    request_ft& operator=(const request_ft&) = delete;
-
-                    request_ft(request_ft&& other) noexcept
-                    : m_req{ std::exchange(other.m_req, nullptr) } 
-                    {}
-
-                    request_ft& operator=(request_ft&& other) noexcept
-                    {
-                        if (m_req) destroy();
-                        m_req = std::exchange(other.m_req, nullptr);
-                    }
-
-                    ~request_ft()
-                    {
-                        if (m_req) destroy();
-                    }
-
-                    void destroy()
-                    {
-                        void* ucx_ptr = m_req->m_ucx_ptr;
-                        m_req->m_send_worker->m_parallel_context->thread_primitives().critical(
-                            [ucx_ptr]()
-                            {
-                                //request_init(ucx_ptr);
-                                ucp_request_free(ucx_ptr);
-                            }
-                        );
-                    }
-
-                    bool test()
-                    {
-                        if (!m_req) return true;
-
-                        ucp_worker_progress(m_req->m_send_worker->get());
-                        ucp_worker_progress(m_req->m_send_worker->get());
-                        ucp_worker_progress(m_req->m_send_worker->get());
-                        
-                        return m_req->m_send_worker->m_parallel_context->thread_primitives().critical(
-                            [this]()
-                            {
-                                ucp_worker_progress(m_req->m_recv_worker->get());
-                                ucp_worker_progress(m_req->m_recv_worker->get());
-
-			                    // check request status
-                                // TODO check whether ucp_request_check_status has to be locked also:
-                                // it does access the worker!
-                                // TODO are we allowed to call this?
-                                // m_req might be a send request submitted on another thread, and hence might access
-                                // the other threads's send worker... 
-			                    if (UCS_INPROGRESS != ucp_request_check_status(m_req->m_ucx_ptr)) 
-                                {
-                                    //request_init(m_req->m_ucx_ptr);
-				                    ucp_request_free(m_req->m_ucx_ptr);
-                                    m_req = nullptr;
-                                    return true;
-                                }
-                                else
-                                    return false;
-                            }
-                        );
-                    }
-                    
-                    void wait()
-                    {
-                        if (!m_req) return;
-                        while (!test());
-                    }
-                    
-                    bool cancel()
-                    {
-                        if (!m_req) return true;
-
-                        // TODO at this time, send requests cannot be canceled
-                        // in UCX (1.7.0rc1)
-                        // https://github.com/openucx/ucx/issues/1162
-                        // 
-                        // TODO the below is only correct for recv requests,
-                        // or for send requests under the assumption that 
-                        // the requests cannot be moved between threads.
-                        // 
-                        // For the send worker we do not use locks, hence
-                        // if request is canceled on another thread, it might
-                        // clash with another send being submitted by the owner
-                        // of ucp_worker
-
-                        if (m_req->m_kind == request_kind::send) return false;
-
-                        m_req->m_send_worker->m_parallel_context->thread_primitives().critical(
-                            [this]()
-                            {
-                                ucp_request_cancel(m_req->m_recv_woker->get(), m_req->m_ucx_ptr);
-                            }
-                        );
-			            // wait for the request to either complete, or be canceled
-                        wait();
-                        return true;
-                    }
-                };
-
 
                 template<typename ThreadPrimitives>
                 struct communicator
@@ -174,18 +35,23 @@ namespace gridtools {
                     using rank_type              = endpoint_t::rank_type;
                     using tag_type               = int;
                     using request                = request_ft<ThreadPrimitives>;
+                    template<typename T>
+                    using future                 = future_t<T,ThreadPrimitives>;
+                    // needed for now for high-level API
+                    using address_type           = rank_type;
 
                     worker_type*  m_recv_worker;
                     worker_type*  m_send_worker;
 
                     rank_type rank() const noexcept { return m_send_worker->rank(); }
                     rank_type size() const noexcept { return m_send_worker->size(); }
+                    address_type address() const { return rank(); }
 
                     static void empty_send_callback(void *, ucs_status_t) {}
                     static void empty_recv_callback(void *, ucs_status_t, ucp_tag_recv_info_t*) {}
 
                     template <typename MsgType>
-                    [[nodiscard]] request send(const MsgType &msg, rank_type dst, tag_type tag)
+                    [[nodiscard]] future<void> send(const MsgType &msg, rank_type dst, tag_type tag)
                     {
                         const auto& ep = m_send_worker->connect(dst);
                         const auto stag = ((std::uint_fast64_t)tag << 32) | 
@@ -215,7 +81,7 @@ namespace gridtools {
                     }
 		
                     template <typename MsgType>
-                    [[nodiscard]] request recv(MsgType &msg, rank_type src, tag_type tag)
+                    [[nodiscard]] future<void> recv(MsgType &msg, rank_type src, tag_type tag)
                     {
                         const auto rtag = ((std::uint_fast64_t)tag << 32) | 
                                            (std::uint_fast64_t)(src);
