+++ include/ghex/transport_layer/ucx/request2.hpp
@@ -1,0 +1,168 @@
+/* 
+ * GridTools
+ * 
+ * Copyright (c) 2014-2019, ETH Zurich
+ * All rights reserved.
+ * 
+ * Please, refer to the LICENSE file in the root directory.
+ * SPDX-License-Identifier: BSD-3-Clause
+ * 
+ */
+#ifndef INCLUDED_GHEX_TL_UCX_REQUEST_HPP
+#define INCLUDED_GHEX_TL_UCX_REQUEST_HPP
+
+#include "../context.hpp"
+#include "../ucx3/worker.hpp"
+#include "../../threads/atomic/primitives.hpp"
+
+namespace gridtools{
+    namespace ghex {
+        namespace tl {
+            namespace ucx {
+
+                enum class request_kind { send, recv };
+
+                template<typename ThreadPrimitives>
+                struct request_data_ft
+                {
+                    using worker_type            = worker_t<ThreadPrimitives>;
+
+                    void*        m_ucx_ptr;
+                    worker_type* m_recv_worker;
+                    worker_type* m_send_worker;
+                    request_kind m_kind;
+
+                    static constexpr std::uintptr_t mask = ~(alignof(request_data_ft)-1u);
+                    template<typename... Args>
+                    static request_data_ft* construct(void* ptr, Args&& ...args)
+                    {
+                        // align pointer    
+                        auto a_ptr = reinterpret_cast<request_data_ft*>
+                            ((reinterpret_cast<std::uintptr_t>((unsigned char*)ptr) + alignof(request_data_ft)-1) & mask);
+                        new(a_ptr) request_data_ft{ptr,std::forward<Args>(args)...};
+                        return a_ptr;
+                    }
+                };
+                
+                using request_data_size = std::integral_constant<std::size_t,
+                    sizeof(request_data_ft<::gridtools::ghex::threads::atomic::primitives>) +
+                    alignof(request_data_ft<::gridtools::ghex::threads::atomic::primitives>)>;
+                
+                void request_init(void *req) { std::memset(req, 0, request_data_size::value); }
+
+                template<typename ThreadPrimitives>
+                struct request_ft
+                {
+                    using data_type = request_data_ft<ThreadPrimitives>;
+
+                    data_type* m_req = nullptr;
+
+                    request_ft() = default;
+                    request_ft(data_type* ptr) noexcept : m_req{ptr} {}
+                    request_ft(const request_ft&) = delete;
+                    request_ft& operator=(const request_ft&) = delete;
+
+                    request_ft(request_ft&& other) noexcept
+                    : m_req{ std::exchange(other.m_req, nullptr) } 
+                    {}
+
+                    request_ft& operator=(request_ft&& other) noexcept
+                    {
+                        if (m_req) destroy();
+                        m_req = std::exchange(other.m_req, nullptr);
+                    }
+
+                    ~request_ft()
+                    {
+                        if (m_req) destroy();
+                    }
+
+                    void destroy()
+                    {
+                        void* ucx_ptr = m_req->m_ucx_ptr;
+                        m_req->m_send_worker->m_parallel_context->thread_primitives().critical(
+                            [ucx_ptr]()
+                            {
+                                //request_init(ucx_ptr);
+                                ucp_request_free(ucx_ptr);
+                            }
+                        );
+                    }
+
+                    bool test()
+                    {
+                        if (!m_req) return true;
+
+                        ucp_worker_progress(m_req->m_send_worker->get());
+                        ucp_worker_progress(m_req->m_send_worker->get());
+                        ucp_worker_progress(m_req->m_send_worker->get());
+                        
+                        return m_req->m_send_worker->m_parallel_context->thread_primitives().critical(
+                            [this]()
+                            {
+                                ucp_worker_progress(m_req->m_recv_worker->get());
+                                ucp_worker_progress(m_req->m_recv_worker->get());
+
+			                    // check request status
+                                // TODO check whether ucp_request_check_status has to be locked also:
+                                // it does access the worker!
+                                // TODO are we allowed to call this?
+                                // m_req might be a send request submitted on another thread, and hence might access
+                                // the other threads's send worker... 
+			                    if (UCS_INPROGRESS != ucp_request_check_status(m_req->m_ucx_ptr)) 
+                                {
+                                    //request_init(m_req->m_ucx_ptr);
+				                    ucp_request_free(m_req->m_ucx_ptr);
+                                    m_req = nullptr;
+                                    return true;
+                                }
+                                else
+                                    return false;
+                            }
+                        );
+                    }
+                    
+                    void wait()
+                    {
+                        if (!m_req) return;
+                        while (!test());
+                    }
+                    
+                    bool cancel()
+                    {
+                        if (!m_req) return true;
+
+                        // TODO at this time, send requests cannot be canceled
+                        // in UCX (1.7.0rc1)
+                        // https://github.com/openucx/ucx/issues/1162
+                        // 
+                        // TODO the below is only correct for recv requests,
+                        // or for send requests under the assumption that 
+                        // the requests cannot be moved between threads.
+                        // 
+                        // For the send worker we do not use locks, hence
+                        // if request is canceled on another thread, it might
+                        // clash with another send being submitted by the owner
+                        // of ucp_worker
+
+                        if (m_req->m_kind == request_kind::send) return false;
+
+                        m_req->m_send_worker->m_parallel_context->thread_primitives().critical(
+                            [this]()
+                            {
+                                ucp_request_cancel(m_req->m_recv_woker->get(), m_req->m_ucx_ptr);
+                            }
+                        );
+			            // wait for the request to either complete, or be canceled
+                        wait();
+                        return true;
+                    }
+                };
+
+            } // namespace ucx
+        } // namespace tl
+    } // namespace ghex
+} // namespace gridtools
+
+#endif /* INCLUDED_GHEX_TL_UCX_REQUEST_HPP */
+
